pyproject.toml
[project]
name = "gst-itr-bot"
version = "0.1.0"
description = "WhatsApp-based GST + ITR filing assistant with OCR, STT, and sandbox integrations"
requires-python = ">=3.10"

dependencies = [
    # --- Web framework / API ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # --- DB / ORM ---
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",

    # --- Cache / queues ---
    "redis>=5.0.0",

    # --- Config / utils / logging ---
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.3.0",

    # --- Templates / admin UI ---
    "jinja2>=3.1.0",

    # --- HTTP clients ---
    "requests>=2.31.0",
    "aiohttp>=3.9.0",

    # --- OCR & PDF tools ---
    "Pillow>=10.0.0",        # image handling
    "PyMuPDF>=1.24.0",       # PDF reading
    "reportlab>=4.0.0",      # PDF generation / invoice summary export

    # --- Cloud OCR (Google Vision) ---
    "google-cloud-vision>=3.11.0",
    "google-cloud-storage>=2.14.0",

    # --- Optional: PaddleOCR (Linux/Intel only; usually NOT for M1) ---
    # "paddlepaddle>=2.6.0",
    # "paddleocr>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "UP"]
ignore = []